Register service actions via async_setup
